LPS-54043 Move
