+++ modules/apps/service-access-control-profile/service-access-control-profile-service/src/com/liferay/service/access/control/profile/ServiceAccessControlProfileManagerImpl.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.service.access.control.profile.service.security;
+package com.liferay.service.access.control.profile;
 
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.security.ac.ServiceAccessControlProfile;
-import com.liferay.portal.kernel.security.ac.ServiceAccessControlProfileManager;
+import com.liferay.portal.kernel.security.access.control.profile.ServiceAccessControlProfile;
+import com.liferay.portal.kernel.security.access.control.profile.ServiceAccessControlProfileManager;
 import com.liferay.service.access.control.profile.model.SACPEntry;
 import com.liferay.service.access.control.profile.service.SACPEntryService;
