+++ portal-service/src/com/liferay/portal/kernel/security/access/control/profile/ServiceAccessControlProfileThreadLocal.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.security.ac;
+package com.liferay.portal.kernel.security.access.control.profile;
 
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
